+++ lib/cuckoo/common/abstracts.py
@@ -28,6 +28,7 @@ from lib.cuckoo.common.exceptions import (
     CuckooOperationalError,
     CuckooReportError,
 )
+from lib.cuckoo.common.integrations.mitre import mitre_load
 from lib.cuckoo.common.objects import Dictionary
 from lib.cuckoo.common.url_validate import url as url_validator
 from lib.cuckoo.common.utils import create_folder, get_memdump_path, load_categories
@@ -55,32 +56,8 @@ except ImportError:
 
 repconf = Config("reporting")
 _, categories_need_VM = load_categories()
-HAVE_MITRE = False
 
-if repconf.mitre.enabled:
-    try:
-        from pyattck import Attck
-        from pyattck.utils.version import __version_info__ as pyattck_version
-
-        if pyattck_version >= (4, 1, 1) and pyattck_version <= (5, 2, 0):
-            mitre = Attck(
-                nested_subtechniques=True,
-                use_config=True,
-                save_config=True,
-                config_file_path=os.path.join(CUCKOO_ROOT, "data", "mitre", "config.yml"),
-                data_path=os.path.join(CUCKOO_ROOT, "data", "mitre"),
-                enterprise_attck_json=os.path.join(CUCKOO_ROOT, "data", "mitre", "enterprise_attck_json.json"),
-                pre_attck_json=os.path.join(CUCKOO_ROOT, "data", "mitre", "pre_attck_json.json"),
-                mobile_attck_json=os.path.join(CUCKOO_ROOT, "data", "mitre", "mobile_attck_json.json"),
-                ics_attck_json=os.path.join(CUCKOO_ROOT, "data", "mitre", "ics_attck_json.json"),
-                nist_controls_json=os.path.join(CUCKOO_ROOT, "data", "mitre", "nist_controls_json.json"),
-                generated_attck_json=os.path.join(CUCKOO_ROOT, "data", "mitre", "generated_attck_json.json"),
-                generated_nist_json=os.path.join(CUCKOO_ROOT, "data", "mitre", "generated_nist_json.json"),
-            )
-            HAVE_MITRE = True
-
-    except ImportError:
-        print("Missed pyattck dependency: check requirements.txt for exact pyattck version")
+mitre, HAVE_MITRE, _ = mitre_load(repconf.mitre.enabled)
 
 log = logging.getLogger(__name__)
 cfg = Config()
@@ -1528,8 +1505,11 @@ class Signature:
         return res
 
     def mark_call(self, *args, **kwargs):
-        """Mark the current call as explanation as to why this signature
-        matched."""
+        """Mark the current call as explanation as to why this signature matched."""
+
+        if not cfg.cuckoo.apicall_details:
+            return
+
         mark = {
             "type": "call",
             "pid": self.pid,