+++ utils/community.py
@@ -22,13 +22,14 @@ sys.path.append(os.path.join(os.path.abspath(os.path.dirname(__file__)), ".."))
 
 import lib.cuckoo.common.colors as colors
 from lib.cuckoo.common.constants import CUCKOO_ROOT
+from lib.cuckoo.common.integrations.mitre import mitre_update
+from lib.cuckoo.common.path_utils import path_exists, path_mkdir
 
 blocklist = {}
 if os.path.exists(os.path.join(CUCKOO_ROOT, "utils", "community_blocklist.py")):
     from utils.community_blocklist import blocklist
 
 log = logging.getLogger(__name__)
-URL = "https://github.com/kevoreilly/community/archive/{0}.tar.gz"
 
 
 def flare_capa():
@@ -62,48 +63,25 @@ def flare_capa():
         print(e)
 
 
-def mitre():
-    """Urls might change, for proper urls see https://github.com/swimlane/pyattck"""
-    try:
-        from pyattck import Attck
-    except ImportError:
-        print("Missed dependency: install pyattck library, see requirements for proper version")
-        return
-
-    mitre = Attck(
-        nested_subtechniques=True,
-        use_config=False,
-        save_config=False,
-        config_file_path=os.path.join(CUCKOO_ROOT, "data", "mitre", "config.yml"),
-        data_path=os.path.join(CUCKOO_ROOT, "data", "mitre"),
-        enterprise_attck_json="https://raw.githubusercontent.com/mitre/cti/master/enterprise-attack/enterprise-attack.json",
-        pre_attck_json="https://raw.githubusercontent.com/mitre/cti/master/pre-attack/pre-attack.json",
-        mobile_attck_json="https://raw.githubusercontent.com/mitre/cti/master/mobile-attack/mobile-attack.json",
-        ics_attck_json="https://raw.githubusercontent.com/mitre/cti/master/ics-attack/ics-attack.json",
-        nist_controls_json="https://raw.githubusercontent.com/center-for-threat-informed-defense/attack-control-framework-mappings/master/frameworks/ATT%26CK-v9.0/nist800-53-r4/stix/nist800-53-r4-controls.json",
-        generated_attck_json="https://swimlane-pyattck.s3.us-west-2.amazonaws.com/generated_attck_data.json",
-        generated_nist_json="https://swimlane-pyattck.s3.us-west-2.amazonaws.com/attck_to_nist_controls.json",
-    )
-
-    print("[+] Updating MITRE datasets")
-    mitre.update()
-
-
-def install(enabled, force, rewrite, filepath, access_token=None):
-    if filepath and os.path.exists(filepath):
+def install(enabled, force, rewrite, filepath: str = False, access_token=None, proxy=False, url: str = False):
+    if filepath and path_exists(filepath):
         t = tarfile.TarFile.open(filepath, mode="r:gz")
     else:
-        print(f"Downloading modules from {URL}")
+        print(f"Downloading modules from {url}")
         try:
             http = urllib3.PoolManager()
             if access_token is None:
-                data = http.request("GET", URL).data
-            elif "github" in URL:
+                data = http.request("GET", url).data
+            elif "github" in url:
                 data = http.request(
-                    "GET", URL, headers={"Authorization": f"token {access_token}", "User-Agent": "CAPEv2_sandbox"}
+                    "GET", url, headers={"Authorization": f"token {access_token}", "User-Agent": "CAPEv2_sandbox"}
                 ).data
             else:
-                data = http.request("GET", URL, headers={"PRIVATE-TOKEN": access_token}).data
+                data = http.request("GET", url, headers={"PRIVATE-TOKEN": access_token}).data
+
+            if b"Not Found" == data:
+                print("You don't have permissions to access this repo")
+                sys.exit(-1)
             t = tarfile.TarFile.open(fileobj=BytesIO(data), mode="r:gz")
         except Exception as e:
             print("ERROR: Unable to download archive: %s" % e)
@@ -117,6 +95,8 @@ def install(enabled, force, rewrite, filepath, access_token=None):
         "machinery": "modules/machinery",
         "analyzer": "analyzer",
         "data": "data",
+        "integrations": "lib/cuckoo/common/integrations",
+        "mitre": "data/mitre",
     }
 
     members = t.getmembers()
@@ -177,7 +157,6 @@ def install(enabled, force, rewrite, filepath, access_token=None):
 
 
 def main():
-    global URL
 
     parser = argparse.ArgumentParser()
     parser.add_argument("-a", "--all", help="Download everything", action="store_true", required=False)
@@ -188,6 +167,7 @@ def main():
     parser.add_argument("-r", "--reporting", help="Download reporting modules", action="store_true", required=False)
     parser.add_argument("-an", "--analyzer", help="Download analyzer modules/binaries/etc", action="store_true", required=False)
     parser.add_argument("-data", "--data", help="Download data items", action="store_true", required=False)
+    parser.add_argument("-i", "--integrations", help="Download integration items", action="store_true", required=False)
     parser.add_argument(
         "-f", "--force", help="Install files without confirmation", action="store_true", default=False, required=False
     )
@@ -201,6 +181,13 @@ def main():
     )
     parser.add_argument("--mitre", help="Download updated MITRE JSONS", action="store_true", default=False, required=False)
     parser.add_argument(
+        "--mitre-offline",
+        help="Download updated MITRE JSONS from community repo",
+        action="store_true",
+        default=False,
+        required=False,
+    )
+    parser.add_argument(
         "-u", "--url", help="Download community modules from the specified url", action="store", default=None, required=False
     )
     parser.add_argument(
@@ -208,11 +195,10 @@ def main():
     )
     args = parser.parse_args()
 
-    URL = args.url or URL.format(args.branch)
     enabled = []
 
     if args.all:
-        enabled = ["feeds", "processing", "signatures", "reporting", "machinery", "analyzer", "data"]
+        enabled = ["feeds", "processing", "signatures", "reporting", "machinery", "analyzer", "data", "integrations", "mitre"]
         flare_capa()
     else:
         if args.feeds:
@@ -229,6 +215,10 @@ def main():
             enabled.append("analyzer")
         if args.data:
             enabled.append("data")
+        if args.integrations:
+            enabled.append("integrations")
+        if args.mitre_offline:
+            enabled.append("mitre")
 
     if args.capa_rules:
         flare_capa()
@@ -236,7 +226,7 @@ def main():
             return
 
     if args.mitre:
-        mitre()
+        mitre_update()
         if not enabled:
             return
 
@@ -245,7 +235,15 @@ def main():
         parser.print_help()
         return
 
-    install(enabled, args.force, args.rewrite, args.file, args.token)
+    install(
+        enabled,
+        args.force,
+        args.rewrite,
+        args.file,
+        args.token,
+        args.proxy,
+        args.url or f"https://github.com/kevoreilly/community/archive/{args.branch}.tar.gz",
+    )
 
 
 if __name__ == "__main__":