+++ changelog.md
@@ -1,3 +1,24 @@
+### [2.2.2023]
+* Monitor update: Process dump improvements & 'export' option to allow DLL export to be defined by monitor yara signature
+
+### [1.2.2023]
+* Monitor update: Disable spawning WER processes (werfault.exe etc) via RtlReportSilentProcessExit hook
+
+### [30.01.2023]
+* Add `utils/fstab.py` utils which is used by `utils/dist.py` when NFS mode is used.
+    * Check configure NFS in [documentation](https://capev2.readthedocs.io/en/latest/usage/dist.html):
+* Now when you register new server in distributed cluster that uses NFS, it will automatically:
+    * Create worker folder
+    * Add NFS entry to `/etc/fstab`. Ex:
+        * `192.168.1.1:/opt/CAPEv2 /opt/CAPEv2/workers/192.168.1.1 nfs, auto,user,users,nofail,noatime,nolock,intr,tcp,actimeo=1800, 0 0`
+    * Mount folder
+
+### [26.1.2023] Configs
+* Please read [this](https://github.com/kevoreilly/CAPEv2/blob/master/conf/readme.md) to simplify your life with configs managment
+
+### [25.1.2023]
+* Google Cloud Platform (GCP) support in distributed CAPE aka dist.py
+
 ### [5.1.2023]
 * Big duplicated code cleanup. Context: CAPE.py module processing all the files so it calling File(x).get_all() which is pretty heavy.
 * Deprecated standalone modules. They are moved inside of CAPE.py. Data will be under the same keys.